extract state initialization to new module
