+++ src/swarm-tournament/Swarm/Web/Tournament/Validate.hs
@@ -29,6 +29,7 @@ import Swarm.Game.Scenario
 import Swarm.Game.Scenario.Scoring.CodeSize (codeMetricsFromSyntax)
 import Swarm.Game.Scenario.Status (emptyLaunchParams)
 import Swarm.Game.State
+import Swarm.Game.State.Initialize (scenarioToGameState)
 import Swarm.Game.State.Runtime (initGameStateConfig, initScenarioInputs)
 import Swarm.Game.State.Substate (initState, seed)
 import Swarm.Game.Step.Validate (playUntilWin)
