+++ test/bench/Benchmark.hs
@@ -21,7 +21,8 @@ import Swarm.Game.Location
 import Swarm.Game.Robot (TRobot, mkRobot)
 import Swarm.Game.Robot.Walk (emptyExceptions)
 import Swarm.Game.Scenario (loadStandaloneScenario)
-import Swarm.Game.State (GameState, creativeMode, landscape, pureScenarioToGameState, zoomRobots)
+import Swarm.Game.State (GameState, creativeMode, landscape, zoomRobots)
+import Swarm.Game.State.Initialize (pureScenarioToGameState)
 import Swarm.Game.State.Landscape (multiWorld)
 import Swarm.Game.State.Robot (addTRobot)
 import Swarm.Game.State.Runtime (initRuntimeState, stdGameConfigInputs)
